distribute assemblers over files for respective types
